@@ -13,11 +13,11 @@
     }
   ],
   "require": {
-    "php": ">=7.2",
+    "php": ">=7.4",
     "ext-pdo": "*",
     "ext-json": "*",
     "ext-openssl": "*",
-    "twig/twig":"~3.0",
+    "twig/twig":"~3.4.0",
     "phpunit/phpunit": "^8.0",
     "predis/predis":"^1.1",
     "auth0/auth0-php": "^7.4",
