@@ -30,8 +30,8 @@ interface IConnection extends SplSubject
 	public static function get();
 	public function getDriver();
 	public function getLastQuery();
-	public function setLastQuery($lastQuery);
-	public function addQueryToLog($query);
+	public function setLastQuery(string $lastQuery);
+	public function addQueryToLog($query, array $params = []);
 	public function exec($statement);
 	public function query($statement);
 	public function prepare($statement, $driver_options = []);
